Basic Scanner UI Finishes + Button Fixes
This is just the basic UI for the scanner. I will be researching how to implement a camera view into a scene soon, as well as AR image tracking. This is now the start of the actual coding for this app and I'm EXCITED!

I also fixed the scanner button on every page with a bottom navbar that shows the button for the scanner page. They now all should go to the scanner page instead of the home page.